+++ changelog.md
@@ -6,4 +6,8 @@
   * CN, specialisation of CollectErrors to NumErrors
   * numerical partial operators (eg division) return a CN type
   * instances for Data.Complex
-  
+
+* v 0.2.0.1
+  * fix compilation bug in test suite
+  * minor doc improvements
+  * fix Complex instances of error-throwing division (/!)